Replace unsigned int with size_t where value indicates a count.

+++ include/skiplist.h
@@ -44,16 +44,16 @@ skiplist*	skiplist_new(dict_compare_func cmp_func, dict_delete_func del_func,
 						 unsigned max_link);
 dict*		skiplist_dict_new(dict_compare_func cmp_func,
 							  dict_delete_func del_func, unsigned max_link);
-unsigned	skiplist_free(skiplist *list);
+size_t		skiplist_free(skiplist *list);
 
 int			skiplist_insert(skiplist *list, void *key, void *datum,
 							int overwrite);
 int			skiplist_probe(skiplist *list, void *key, void **datum);
 void*		skiplist_search(skiplist *list, const void *key);
 int			skiplist_remove(skiplist *list, const void *key);
-unsigned	skiplist_clear(skiplist *list);
-unsigned	skiplist_traverse(skiplist *list, dict_visit_func visit);
-unsigned	skiplist_count(const skiplist *list);
+size_t		skiplist_clear(skiplist *list);
+size_t		skiplist_traverse(skiplist *list, dict_visit_func visit);
+size_t		skiplist_count(const skiplist *list);
 
 typedef struct skiplist_itor
 			skiplist_itor;
@@ -67,8 +67,8 @@ int			skiplist_itor_valid(const skiplist_itor *itor);
 void		skiplist_itor_invalidate(skiplist_itor *itor);
 int			skiplist_itor_next(skiplist_itor *itor);
 int			skiplist_itor_prev(skiplist_itor *itor);
-int			skiplist_itor_nextn(skiplist_itor *itor, unsigned count);
-int			skiplist_itor_prevn(skiplist_itor *itor, unsigned count);
+int			skiplist_itor_nextn(skiplist_itor *itor, size_t count);
+int			skiplist_itor_prevn(skiplist_itor *itor, size_t count);
 int			skiplist_itor_first(skiplist_itor *itor);
 int			skiplist_itor_last(skiplist_itor *itor);
 int			skiplist_itor_search(skiplist_itor *itor, const void *key);
